requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
factory_boy==3.3.1
Faker==33.1.0
mccabe==0.7.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
